pyproject.toml
[project]
name = "taskmanagementapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.10",
    "python-dotenv>=1.1.1",
]
